refactor: separate animated top snack bar
