+++ lib/common/action/notifications.dart
@@ -1,99 +1,18 @@
 import 'package:flutter/material.dart';
+import 'package:remembeer/common/widget/animated_top_snack_bar.dart';
 
 void _showTopSnackBar(BuildContext context, String message) {
   final overlay = Overlay.of(context);
   late OverlayEntry overlayEntry;
 
   overlayEntry = OverlayEntry(
-    builder: (context) => _AnimatedTopSnackBar(
+    builder: (context) => AnimatedTopSnackBar(
       message: message,
       onDismissed: () => overlayEntry.remove(),
     ),
   );
 
   overlay.insert(overlayEntry);
-}
-
-class _AnimatedTopSnackBar extends StatefulWidget {
-  final String message;
-  final VoidCallback onDismissed;
-
-  const _AnimatedTopSnackBar({
-    required this.message,
-    required this.onDismissed,
-  });
-
-  @override
-  State<_AnimatedTopSnackBar> createState() => _AnimatedTopSnackBarState();
-}
-
-class _AnimatedTopSnackBarState extends State<_AnimatedTopSnackBar>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<Offset> _slideAnimation;
-  late Animation<double> _fadeAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
-
-    _slideAnimation = Tween<Offset>(
-      begin: const Offset(0, -1),
-      end: Offset.zero,
-    ).animate(CurvedAnimation(parent: _controller, curve: Curves.easeOut));
-
-    _fadeAnimation = Tween<double>(
-      begin: 0,
-      end: 1,
-    ).animate(CurvedAnimation(parent: _controller, curve: Curves.easeOut));
-
-    _controller.forward();
-
-    Future.delayed(const Duration(seconds: 2), () {
-      if (mounted) {
-        _controller.reverse().then((_) => widget.onDismissed());
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      top: MediaQuery.of(context).padding.top + 8,
-      left: 16,
-      right: 16,
-      child: SlideTransition(
-        position: _slideAnimation,
-        child: FadeTransition(
-          opacity: _fadeAnimation,
-          child: Material(
-            elevation: 6,
-            borderRadius: BorderRadius.circular(8),
-            color: Theme.of(context).colorScheme.inverseSurface,
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-              child: Text(
-                widget.message,
-                style: TextStyle(
-                  color: Theme.of(context).colorScheme.onInverseSurface,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
 }
 
 void showDefaultDrinkAdded(BuildContext context) {
